Refactor auth scontext imports
