+++ src/screens/ChangeLang/index.js
@@ -6,7 +6,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { Button } from 'react-native-elements';
 import { LanguageContext } from '../../state/LanguageContext';
 import { StyledView } from '../../styles/shared/StyledView';
-import { Context as AuthContext } from '../../state/AuthContext';
+import { AuthContext } from 'state/';
 import { styles } from './ChangeLang.styles';
 
 export default function ChangeLanguage({ navigation }) {
