+++ src/state/index.js
@@ -1,0 +1,4 @@
+import { Context as AuthContext } from './AuthContext.js';
+import { Provider as AuthProvider } from './AuthContext';
+
+export { AuthContext, AuthProvider };
